+++ app/Loader.scala
@@ -4,7 +4,7 @@ import play.api.ApplicationLoader.Context
 import play.api.{Application, ApplicationLoader, Configuration, LoggerConfigurator, Mode}
 import switchboard.{SwitchboardConfiguration, Lifecycle => SwitchboardLifecycle}
 import conf.ApplicationConfiguration
-import services.editions.publishing.events.IssuePublishEventsListener
+import services.editions.publishing.events.PublishEventsListener
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
@@ -50,7 +50,7 @@ class Loader extends ApplicationLoader {
     )
     new LogStashConfig(components.config)
 
-    IssuePublishEventsListener.apply(config).start
+    PublishEventsListener.apply(config).start
 
     components.application
   }
